+++ lib/app/view_app.dart
@@ -1,2 +1,4 @@
 //
 export 'view/app_view.dart';
+
+export 'view/splash_screen.dart';